[FLINK-2105] Extract abstract superclass, interface from MergeMatchIterators, KeyGroupedIterators
